Move generateTranscript to utils and change file ending
